requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.2
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
httpie==2.0.0
idna==2.8
itypes==1.1.0
Jinja2==2.11.0
Markdown==3.1.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2==2.8.4
Pygments==2.5.2
PyJWT==1.7.1
python-decouple==3.3
pytz==2019.3
requests==2.22.0
simplejson==3.17.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.7
whitenoise==5.0.1
pillow==7.0.0
